rss_pipeline/requirements.txt
psycopg2-binary
pandas
python-dotenv
pylint
pytest
openapi
requests